Move theme setting into seperate file
